os2ds/src/data/all/requirements.txt
lorem
Pillow